Misra fix

+++ src/software/firmware/srcs/pressure_valve.cpp
@@ -48,4 +48,4 @@ void PressureValve::open() { command = openApertureAngle; }
 
 void PressureValve::close() { command = closeApertureAngle; }
 
-void PressureValve::open(int32_t p_command) { command = p_command; }
+void PressureValve::open(uint16_t p_command) { command = p_command; }
